+++ cmd/backfillUrns.go
@@ -4,14 +4,21 @@ import (
 	"fmt"
 	"sync"
 
+	"github.com/ethereum/go-ethereum/ethclient"
+	"github.com/ethereum/go-ethereum/rpc"
 	"github.com/makerdao/vdb-mcd-transformers/backfill"
 	"github.com/makerdao/vdb-mcd-transformers/backfill/fork"
 	"github.com/makerdao/vdb-mcd-transformers/backfill/frob"
 	"github.com/makerdao/vdb-mcd-transformers/backfill/grab"
 	"github.com/makerdao/vdb-mcd-transformers/backfill/repository"
 	"github.com/makerdao/vdb-mcd-transformers/backfill/shared"
+	"github.com/makerdao/vulcanizedb/pkg/config"
 	"github.com/makerdao/vulcanizedb/pkg/core"
 	"github.com/makerdao/vulcanizedb/pkg/datastore/postgres/repositories"
+	"github.com/makerdao/vulcanizedb/pkg/eth"
+	"github.com/makerdao/vulcanizedb/pkg/eth/client"
+	"github.com/makerdao/vulcanizedb/pkg/eth/converters"
+	"github.com/makerdao/vulcanizedb/pkg/eth/node"
 	"github.com/makerdao/vulcanizedb/utils"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
@@ -37,6 +44,7 @@ var backfillUrnsCmd = &cobra.Command{
 	Long: `Fetch diffs when events indicate the state of an Urn changed at a given block.
 Optionally pass a starting block number to backfill since a given block.
 Optionally pass events to watch (fork, frob, grab) to backfill based off of certain events.`,
+	PreRun: setViperConfigs,
 	Run: func(cmd *cobra.Command, args []string) {
 		err := backfillUrns()
 		if err != nil {
@@ -65,6 +73,18 @@ func init() {
 	viper.BindPFlag("database.user", backfillUrnsCmd.Flags().Lookup("database-user"))
 	viper.BindPFlag("database.password", backfillUrnsCmd.Flags().Lookup("database-password"))
 	viper.BindPFlag("client.ipcPath", backfillUrnsCmd.Flags().Lookup("client-ipcPath"))
+}
+
+func setViperConfigs(cmd *cobra.Command, args []string) {
+	ipc = viper.GetString("client.ipcpath")
+	databaseConfig = config.Database{
+		Name:     viper.GetString("database.name"),
+		Hostname: viper.GetString("database.hostname"),
+		Port:     viper.GetInt("database.port"),
+		User:     viper.GetString("database.user"),
+		Password: viper.GetString("database.password"),
+	}
+	viper.Set("database.config", databaseConfig)
 }
 
 func backfillUrns() error {
@@ -114,4 +134,24 @@ func backFillEvents(backFiller backfill.BackFiller, startingBlock int, errs chan
 		errs <- err
 	}
 	return
+}
+
+func getBlockChain() *eth.BlockChain {
+	rpcClient, ethClient := getClients()
+	vdbEthClient := client.NewEthClient(ethClient)
+	vdbNode := node.MakeNode(rpcClient)
+	transactionConverter := converters.NewTransactionConverter(ethClient)
+	return eth.NewBlockChain(vdbEthClient, rpcClient, vdbNode, transactionConverter)
+}
+
+func getClients() (client.RpcClient, *ethclient.Client) {
+	rawRpcClient, err := rpc.Dial(ipc)
+
+	if err != nil {
+		logrus.Fatal(err)
+	}
+	rpcClient := client.NewRpcClient(rawRpcClient, ipc)
+	ethClient := ethclient.NewClient(rawRpcClient)
+
+	return rpcClient, ethClient
 }
